project/app/rest_framework/serializers.py
from rest_framework_json_api import serializers
from ..models import Wallet, Transaction


class TransactionSerializer(serializers.ModelSerializer):
    class Meta:
        model = Transaction
        fields = ['id', 'wallet', 'txid', 'amount']


class WalletSerializer(serializers.ModelSerializer):
    class Meta:
        model = Wallet
        fields = ['id', 'label', 'balance']
        read_only_fields = ['balance']
