project/pyproject.toml
[tool.poetry]
name = "broker-test-task"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.7"
djangorestframework = "^3.15.2"
mysqlclient = "^2.2.4"
djangorestframework-jsonapi = "^7.0.2"
factory-boy = "^3.3.0"
django-filter = "^24.2"
gunicorn = "^22.0.0"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
